FeedBackLogs now in code, EmpManagement updated

+++ Payroll_ProtoVB/Payroll_ProtoVB/EmpManagment.vb
@@ -9,6 +9,8 @@ Public Class EmpManagment
     Dim Uname As String
     Dim PWord As String
     Dim VryPword As String
+    Dim feedBackLogs As New FeedbackLogsTableAdapter
+    Dim futureTA As New EmployeeFutureTableAdapter
 
     Public Sub New(nRow As Integer)
         InitializeComponent()
@@ -22,6 +24,13 @@ Public Class EmpManagment
     Private Sub EmpManagment_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         'TODO: This line of code loads data into the 'Primary.Employee' table. You can move, or remove it, as needed.
         Me.employTA.FillBySelectedRow(Me.Primary.Employee, Row)
+        If (PayTypeTxtBox.Text = "True") Then
+            SalariedGroupBox.Visible = True
+            HourlyGroupBox.Visible = False
+        Else
+            SalariedGroupBox.Visible = False
+            HourlyGroupBox.Visible = True
+        End If
 
         ' If MaritalStatusTxt.Text = "True" Then
         'MaritalStatusTxt.Text = "Married"
@@ -37,11 +46,33 @@ Public Class EmpManagment
     End Sub
 
     Private Sub SearchBtn_Click(sender As Object, e As EventArgs) Handles SearchBtn.Click
-        ' Needs to handle non Integer text
-        Me.employTA.FillByID(Me.Primary.Employee, SearchEmpTxt.Text)
+        feedBackLogs.CountPlusOne("EmpManagement", "Search Button")
+
+        Try
+            Me.employTA.FillByID(Me.Primary.Employee, SearchEmpTxt.Text)
+            If (PayTypeTxtBox.Text = "True") Then
+                SalariedGroupBox.Visible = True
+                HourlyGroupBox.Visible = False
+            Else
+                SalariedGroupBox.Visible = False
+                HourlyGroupBox.Visible = True
+            End If
+        Catch
+            Dim MSG, style, title, response, MyString
+            MSG = "Not a valid employee ID, please try again."
+            title = "Input Error"
+            style = vbOKOnly + vbDefaultButton1
+            response = MsgBox(MSG, style, title)
+            If response = vbOKOnly Then
+                MyString = "OK"
+            End If
+        End Try
+
     End Sub
 
     Private Sub ConLoginBtn_Click(sender As Object, e As EventArgs) Handles ConLoginBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Continue Button")
+
         Uname = UnameTxt.Text
         PWord = PwordTxt.Text
 
@@ -74,9 +105,19 @@ Public Class EmpManagment
             End If
         End If
 
+        If (PayTypeTxtBox.Text = "True") Then
+            SalariedGroupBox.Visible = True
+            HourlyGroupBox.Visible = False
+        Else
+            SalariedGroupBox.Visible = False
+            HourlyGroupBox.Visible = True
+        End If
+
     End Sub
 
     Private Sub removeEmpBtn_Click(sender As Object, e As EventArgs) Handles removeEmpBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Remove Employee Button")
+
         'This button only takes in the ID as a selector to delete the record.
 
         If EmpIDTxt.Text = "" Then
@@ -89,38 +130,53 @@ Public Class EmpManagment
                 MyString = "OK"
             End If
         Else
-            ID = EmpIDTxt.Text
-            employTA.DeleteRow(ID)
-            FnameTxt.Text = ""
-            LnameTxt.Text = ""
-            EmpIDTxt.Text = ""
-            StreetTxt.Text = ""
-            SSN_Txt.Text = ""
-            StreetTxt.Text = ""
-            PayRateTxt.Text = ""
-            SalaryTxt.Text = ""
-            MaritalStatusTxt.Text = ""
-            DependentsTxt.Text = ""
-            PositionTxt.Text = ""
-            paymentTypeCBox.Text = ""
-            PayHrsTxt.Text = ""
-            ID = 0
+            Dim Response As Integer
+            Response = MessageBox.Show("Are you sure you want to delete this record?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+            If Response = vbYes Then
+                ID = EmpIDTxt.Text
+                employTA.DeleteRow(ID)
+                Try
+                    futureTA.DeleteByID(ID)
+                Catch
+                End Try
+                FnameTxt.Text = ""
+                LnameTxt.Text = ""
+                EmpIDTxt.Text = ""
+                StreetTxt.Text = ""
+                SSN_Txt.Text = ""
+                StreetTxt.Text = ""
+                PayRateTxt.Text = ""
+                SalaryTxtBox.Text = ""
+                MaritalStatusTxt.Text = ""
+                DependentsTxt.Text = ""
+                PositionTxt.Text = ""
+                PayTypeTxtBox.Text = ""
+                PayHrsTxt.Text = ""
+                PayfrequencyTextBox.Text = ""
+                ID = 0
+            End If
         End If
     End Sub
 
     Private Sub cancelBtn_Click(sender As Object, e As EventArgs) Handles cancelBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Cancel Button (Right)")
+
         'ATTENTION: This button should navigate to EmpDashboard.vb
         EmpDashboard.Show()
         Me.Close()
     End Sub
 
     Private Sub CancelLoginBtn_Click(sender As Object, e As EventArgs) Handles CancelLoginBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Cancel Button (Left)")
+
         'ATTENTION: This button should navigate to EmpDashboard.vb
         EmpDashboard.Show()
         Me.Close()
     End Sub
 
     Private Sub AddEmpBtn_Click(sender As Object, e As EventArgs) Handles AddEmpBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Add Employee Button")
+
         'ATTENTION: This button should navigate to AddEmployee.vb
         Dim AddEmployee = New AddEmployee("Management")
         AddEmployee.Show()
@@ -128,12 +184,16 @@ Public Class EmpManagment
     End Sub
 
     Private Sub editEmpBtn_Click(sender As Object, e As EventArgs) Handles editEmpBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Update Button")
+
         Me.EmployeeBindingSource.EndEdit()
         'Me.TableAdapterManager1.UpdateAll(Me.Primary)
         employTA.Adapter.Update(Me.Primary)
     End Sub
 
     Private Sub ClearBtn_Click(sender As Object, e As EventArgs) Handles ClearBtn.Click
+        feedBackLogs.CountPlusOne("EmpManagement", "Clear Selection Button")
+
         FnameTxt.Text = ""
         LnameTxt.Text = ""
         EmpIDTxt.Text = ""
@@ -141,11 +201,12 @@ Public Class EmpManagment
         SSN_Txt.Text = ""
         StreetTxt.Text = ""
         PayRateTxt.Text = ""
-        SalaryTxt.Text = ""
+        SalaryTxtBox.Text = ""
         MaritalStatusTxt.Text = ""
         DependentsTxt.Text = ""
         PositionTxt.Text = ""
-        paymentTypeCBox.Text = ""
+        PayTypeTxtBox.Text = ""
+        PayfrequencyTextBox.Text = ""
         PayHrsTxt.Text = ""
         ID = 0
     End Sub
